requirements.txt
httpx>=0.25.0
click>=8.1.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
